Finish refactoring route guards to use jslib
