--- src/popup/services/not-private-guard.service.ts
@@ -1,24 +1,0 @@
-import { BrowserApi } from '../../browser/browserApi';
-
-import { Injectable } from '@angular/core';
-import {
-    CanActivate,
-    Router,
-} from '@angular/router';
-
-import { UserService } from 'jslib-common/abstractions/user.service';
-import { VaultTimeoutService } from 'jslib-common/abstractions/vaultTimeout.service';
-
-@Injectable()
-export class NotPrivateGuardService implements CanActivate {
-    constructor(private vaultTimeoutService: VaultTimeoutService, private userService: UserService,
-        private router: Router) { }
-
-    async canActivate() {
-        if (BrowserApi.getBackgroundPage() == null) {
-            this.router.navigate(['private-mode']);
-            return false;
-        }
-        return true;
-    }
-}
