+++ src/popup/services/services.module.ts
@@ -6,11 +6,12 @@ import {
 
 import { ToasterModule } from 'angular2-toaster';
 
+import { LockGuardService } from './lock-guard.service';
 import { LaunchGuardService } from './launch-guard.service';
+import { UnauthGuardService } from './unauth-guard.service';
 
 import { AuthGuardService } from 'jslib-angular/services/auth-guard.service';
 import { BroadcasterService } from 'jslib-angular/services/broadcaster.service';
-import { LockGuardService } from 'jslib-angular/services/lock-guard.service';
 import { ValidationService } from 'jslib-angular/services/validation.service';
 
 import { BrowserApi } from '../../browser/browserApi';
@@ -115,8 +116,9 @@ export function initFactory(platformUtilsService: PlatformUtilsService, i18nServ
     providers: [
         ValidationService,
         AuthGuardService,
-        LaunchGuardService,
         LockGuardService,
+        LaunchGuardService,
+        UnauthGuardService,
         PopupUtilsService,
         BroadcasterService,
         { provide: MessagingService, useValue: messagingService },
